requirements.txt
google-cloud-aiplatform>=1.66
google-cloud-storage
google-auth
google-auth-oauthlib
python-dotenv
streamlit
pydantic
pypdf
adk
google-genai
